+++ packages/addons-v5/test/unit/commands/addons/wait.unit.test.js
@@ -9,6 +9,7 @@ let _ = require('lodash')
 const lolex = require('lolex')
 const sinon = require('sinon')
 const {expect} = require('chai')
+const theredoc = require('theredoc')
 
 let clock
 const expansionHeaders = {'Accept-Expansion': 'addon_service,plan'}
@@ -31,7 +32,7 @@ describe('addons:wait', () => {
     sandbox.restore()
   })
 
-  context('waiting for an individual add-on', () => {
+  context('waiting for an individual add-on to provision', () => {
     context('when the add-on is provisioned', () => {
       beforeEach(() => {
         this.nockExpectation = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
@@ -168,6 +169,68 @@ describe('addons:wait', () => {
     })
   })
 
+  context('waiting for an individual add-on to deprovision', () => {
+    context('for an add-on that is still deprovisioning', () => {
+      it('waits until the add-on is deprovisioned', () => {
+        const api = nock('https://api.heroku.com')
+          .post('/actions/addons/resolve', {app: null, addon: 'www-redis-2'})
+          .reply(200, [fixtures.addons['www-redis-2']]) // deprovisioning
+          .get('/apps/acme-inc-www/addons/www-redis-2')
+          .reply(200, fixtures.addons['www-redis-2'])
+          .get('/apps/acme-inc-www/addons/www-redis-2')
+          .reply(404, {id: 'not_found', message: 'Not found.'})
+
+        return cmd.run({args: {addon: 'www-redis-2'}, flags: {'wait-interval': '1'}})
+          .then(() => api.done())
+          .then(() => expect(cli.stderr).to.equal('Destroying www-redis-2... done\n'))
+          .then(() => expect(cli.stdout).to.equal(''))
+      })
+
+      it('does NOT notify the user when deprovisioning takes less than 5 seconds', () => {
+        const notifySpy = sandbox.spy(require('@heroku-cli/notifications'), 'notify')
+
+        let deprovisioningAddon = _.clone(fixtures.addons['www-redis-2'])
+        deprovisioningAddon.id = '37f27548-db4a-4ae0-bb48-57125df0ddc2'
+        deprovisioningAddon.name = 'www-redis-3'
+
+        const api = nock('https://api.heroku.com')
+          .post('/actions/addons/resolve', {app: null, addon: 'www-redis-3'})
+          .reply(200, [deprovisioningAddon]) // deprovisioning
+          .get('/apps/acme-inc-www/addons/www-redis-3')
+          .reply(404, {id: 'not_found', message: 'Not found.'})
+
+        return cmd.run({args: {addon: 'www-redis-3'}, flags: {'wait-interval': '1'}})
+          .then(() => api.done())
+          .then(() => expect(notifySpy.called).to.be.false)
+          .then(() => expect(notifySpy.calledOnce).to.be.false)
+      })
+
+      it('notifies the user when provisioning takes longer than 5 seconds', () => {
+        const notifySpy = sandbox.spy(require('@heroku-cli/notifications'), 'notify')
+
+        let deprovisioningAddon = _.clone(fixtures.addons['www-redis-2'])
+        deprovisioningAddon.id = '967dff74-99b4-4fd2-a0f0-79b523d5c0e1'
+        deprovisioningAddon.name = 'www-redis-4'
+
+        const api = nock('https://api.heroku.com')
+          .post('/actions/addons/resolve', {app: null, addon: 'www-redis-4'})
+          .reply(200, [deprovisioningAddon]) // deprovisioning
+          .get('/apps/acme-inc-www/addons/www-redis-4')
+          .reply(200, () => {
+            clock.tick(5000)
+            return deprovisioningAddon
+          })
+          .get('/apps/acme-inc-www/addons/www-redis-4')
+          .reply(404, {id: 'not_found', message: 'Not found.'})
+
+        return cmd.run({args: {addon: 'www-redis-4'}, flags: {'wait-interval': '1'}})
+          .then(() => api.done())
+          .then(() => expect(notifySpy.called).to.be.true)
+          .then(() => expect(notifySpy.calledOnce).to.be.true)
+      })
+    })
+  })
+
   context('waiting for add-ons', () => {
     context('for an app', () => {
       it('waits for addons serially', () => {
@@ -180,9 +243,12 @@ describe('addons:wait', () => {
         let redisAddon = _.clone(fixtures.addons['www-redis'])
         redisAddon.state = 'provisioning'
 
+        let redis2Addon = _.clone(fixtures.addons['www-redis-2'])
+        redis2Addon.state = 'deprovisioning'
+
         let resolverResponse = nock('https://api.heroku.com')
           .get('/apps/acme-inc-www/addons')
-          .reply(200, [ignoredAddon, wwwAddon, redisAddon])
+          .reply(200, [ignoredAddon, wwwAddon, redisAddon, redis2Addon])
 
         let wwwResponse = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
           .get('/apps/acme-inc-www/addons/www-db')
@@ -191,6 +257,10 @@ describe('addons:wait', () => {
         let redisResponse = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
           .get('/apps/acme-inc-www/addons/www-redis')
           .reply(200, redisAddon)
+
+        let redis2Response = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
+          .get('/apps/acme-inc-www/addons/www-redis-2')
+          .reply(200, redis2Addon)
 
         let provisionedWwwAddon = _.clone(fixtures.addons['www-db'])
         provisionedWwwAddon.state = 'provisioned'
@@ -208,13 +278,23 @@ describe('addons:wait', () => {
           .get('/apps/acme-inc-www/addons/www-db')
           .reply(200, provisionedWwwAddon)
 
+        let deprovisionedRedis2Response = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
+          .get('/apps/acme-inc-www/addons/www-redis-2')
+          .reply(404, {id: 'not_found', message: 'Not found.'})
+
         return cmd.run({args: {addon: null}, flags: {'wait-interval': '1'}, app: 'acme-inc-www'})
           .then(() => resolverResponse.done())
           .then(() => redisResponse.done())
           .then(() => wwwResponse.done())
+          .then(() => redis2Response.done())
           .then(() => provisionedRedisResponse.done())
           .then(() => provisionedWwwResponse.done())
-          .then(() => expect(cli.stderr).to.equal('Creating www-db... done\nCreating www-redis... done\n'))
+          .then(() => deprovisionedRedis2Response.done())
+          .then(() => expect(cli.stderr).to.equal(theredoc`
+            Creating www-db... done
+            Creating www-redis... done
+            Destroying www-redis-2... done\n
+          `))
           .then(() => expect(cli.stdout).to.equal('Created www-db as WWW_URL\nCreated www-redis as REDIS_URL\n'))
       })
     })
@@ -230,9 +310,12 @@ describe('addons:wait', () => {
         let redisAddon = _.clone(fixtures.addons['www-redis'])
         redisAddon.state = 'provisioning'
 
+        let redis2Addon = _.clone(fixtures.addons['www-redis-2'])
+        redis2Addon.state = 'deprovisioning'
+
         let resolverResponse = nock('https://api.heroku.com')
           .get('/addons')
-          .reply(200, [ignoredAddon, wwwAddon, redisAddon])
+          .reply(200, [ignoredAddon, wwwAddon, redisAddon, redis2Addon])
 
         let wwwResponse = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
           .get('/apps/acme-inc-www/addons/www-db')
@@ -241,6 +324,10 @@ describe('addons:wait', () => {
         let redisResponse = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
           .get('/apps/acme-inc-www/addons/www-redis')
           .reply(200, redisAddon)
+
+        let redis2Response = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
+          .get('/apps/acme-inc-www/addons/www-redis-2')
+          .reply(200, redis2Addon)
 
         let provisionedWwwAddon = _.clone(fixtures.addons['www-db'])
         provisionedWwwAddon.state = 'provisioned'
@@ -258,13 +345,23 @@ describe('addons:wait', () => {
           .get('/apps/acme-inc-www/addons/www-db')
           .reply(200, provisionedWwwAddon)
 
+        let deprovisionedRedis2Response = nock('https://api.heroku.com', {reqheaders: expansionHeaders})
+          .get('/apps/acme-inc-www/addons/www-redis-2')
+          .reply(404, {id: 'not_found', message: 'Not found.'})
+
         return cmd.run({args: {addon: null}, flags: {'wait-interval': '1'}, app: null})
           .then(() => resolverResponse.done())
           .then(() => redisResponse.done())
           .then(() => wwwResponse.done())
+          .then(() => redis2Response.done())
           .then(() => provisionedRedisResponse.done())
           .then(() => provisionedWwwResponse.done())
-          .then(() => expect(cli.stderr).to.equal('Creating www-db... done\nCreating www-redis... done\n'))
+          .then(() => deprovisionedRedis2Response.done())
+          .then(() => expect(cli.stderr).to.equal(theredoc`
+            Creating www-db... done
+            Creating www-redis... done
+            Destroying www-redis-2... done\n
+          `))
           .then(() => expect(cli.stdout).to.equal('Created www-db as WWW_URL\nCreated www-redis as REDIS_URL\n'))
       })
     })