+++ packages/apps-v5/src/commands/ps/type.js
@@ -3,7 +3,7 @@
 let cli = require('heroku-cli-util')
 const {sortBy, compact} = require('lodash')
 
-const costs = {Free: 0, Eco: 0, Hobby: 7, Basic: 7, 'Standard-1X': 25, 'Standard-2X': 50, 'Performance-M': 250, Performance: 500, 'Performance-L': 500, '1X': 36, '2X': 72, PX: 576}
+const costMonthly = {Free: 0, Eco: 0, Hobby: 7, Basic: 7, 'Standard-1X': 25, 'Standard-2X': 50, 'Performance-M': 250, Performance: 500, 'Performance-L': 500, '1X': 36, '2X': 72, PX: 576, 'Performance-L-RAM': 500, 'Performance-XL': 750, 'Performance-2XL': 1500}
 
 let emptyFormationErr = app => {
   return new Error(`No process types on ${app}.
@@ -14,8 +14,34 @@ https://devcenter.heroku.com/articles/procfile`)
 async function run(context, heroku) {
   let app = context.app
 
+  // will remove this flag once we have
+  // successfully launched larger dyno sizes
+  let isLargerDyno = false
+  const largerDynoFeatureFlag = await heroku.get('/account/features/frontend-larger-dynos')
+
   let parse = async function (args) {
-    if (args.length === 0) return []
+    if (!args || args.length === 0) return []
+
+    // checks for larger dyno sizes
+    // if the feature is not enabled
+    if (!largerDynoFeatureFlag.enabled) {
+      if (args.find(a => a.match(/=/))) {
+        // eslint-disable-next-line array-callback-return
+        compact(args.map(arg => {
+          let match = arg.match(/^([a-zA-Z0-9_]+)=([\w-]+)$/)
+          let size = match[2]
+
+          const largerDynoNames = /^(?!standard-[12]x$)(performance|private|shield)-(l-ram|xl|2xl)$/i
+          isLargerDyno = largerDynoNames.test(size)
+
+          if (isLargerDyno) {
+            const availableDynoSizes = 'eco, basic, standard-1x, standard-2x, performance-m, performance-l, private-s, private-m, private-l, shield-s, shield-m, shield-l'
+            throw new Error(`No such size as ${size}. Use ${availableDynoSizes}.`)
+          }
+        }))
+      }
+    }
+
     let formation = await heroku.get(`/apps/${app}/formation`)
     if (args.find(a => a.match(/=/))) {
       return compact(args.map(arg => {
@@ -32,6 +58,10 @@ Types: ${cli.color.yellow(formation.map(f => f.type).join(', '))}`)
     }
 
     return formation.map(p => ({type: p.type, size: args[0]}))
+  }
+
+  const calculateHourly = function (size) {
+    return costMonthly[size] / 720
   }
 
   let displayFormation = async function () {
@@ -69,7 +99,8 @@ Types: ${cli.color.yellow(formation.map(f => f.type).join(', '))}`)
       type: cli.color.green(d.type),
       size: cli.color.cyan(d.size),
       qty: cli.color.yellow(d.quantity.toString()),
-      'cost/mo': costs[d.size] ? '$' + (costs[d.size] * d.quantity).toString() : '',
+      'cost/hour': calculateHourly(d.size) ? '~$' + (calculateHourly(d.size) * d.quantity).toFixed(3).toString() : '',
+      'max cost/month': costMonthly[d.size] ? '$' + (costMonthly[d.size] * d.quantity).toString() : '',
     }))
 
     if (formation.length === 0) throw emptyFormationErr()
@@ -80,7 +111,8 @@ Types: ${cli.color.yellow(formation.map(f => f.type).join(', '))}`)
         {key: 'type'},
         {key: 'size'},
         {key: 'qty'},
-        {key: 'cost/mo'},
+        {key: 'cost/hour'},
+        {key: 'max cost/month'},
       ],
     })