+++ packages/cli/src/commands/run/detached.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 // tslint:disable:file-name-casing
 import color from '@heroku-cli/color'
-import {Command, flags} from '@heroku-cli/command'
-import {DynoSizeCompletion, ProcessTypeCompletion} from '@heroku-cli/command/lib/completions'
-import {ux} from '@oclif/core'
-
-import Dyno from '../../lib/run/dyno'
-import {buildCommand} from '../../lib/run/helpers'
-import logDisplayer from '../../lib/run/log-displayer'
+import {Command, flags} from '@heroku-cli/command-v9'
+import {DynoSizeCompletion, ProcessTypeCompletion} from '@heroku-cli/command-v9/lib/completions'
+import {CliUx} from '@oclif/core-v1'
+import '@oclif/core-v1/lib/parser'
+import Dyno from '@heroku-cli/plugin-run/lib/lib/dyno'
+import {buildCommand} from '@heroku-cli/plugin-run/lib/lib/helpers'
+import logDisplayer from '@heroku-cli/plugin-run/lib/lib/log-displayer'
 
 export default class RunDetached extends Command {
   static description = 'run a detached dyno, where output is sent to your logs'
@@ -33,7 +33,7 @@ export default class RunDetached extends Command {
     const opts = {
       heroku: this.heroku,
       app: flags.app,
-      command: buildCommand(argv as string[]),
+      command: buildCommand(argv),
       size: flags.size,
       type: flags.type,
       env: flags.env,
@@ -57,7 +57,7 @@ export default class RunDetached extends Command {
       })
     } else {
       // @ts-ignore
-      ux.log(`Run ${color.cmd('heroku logs --app ' + dyno.opts.app + ' --dyno ' + dyno.dyno.name)} to view the output.`)
+      CliUx.ux.log(`Run ${color.cmd('heroku logs --app ' + dyno.opts.app + ' --dyno ' + dyno.dyno.name)} to view the output.`)
     }
   }
 }